created a Navbar
